### 2021-06-29 #3432 @whedon
[PRE REVIEW]: NiTransforms: A Python tool to read, represent, manipulate, and apply n-dimensional spatial transforms
**Submitting author:** @mgxd (<a href="http://orcid.org/0000-0002-7252-7771">Mathias Goncalves</a>)
**Repository:** <a href="https://github.com/poldracklab/nitransforms" target ="_blank">https://github.com/poldracklab/nitransforms</a>
**Version:** 21.0.0
