### 2021-07-02 bf249ca
FIX: Implement AFNI's deoblique operations
Finalize the interpretation in I/O of AFNI's linear matrices when reference
or moving data are oblique w.r.t. the canonical axes.

Resolves: #45.
