### 2021-07-06 #3432 @osorensen
@PeerHerholz, @snastase, would any of you be interested in reviewing this paper for the Journal of Open Source Software?
