### 2021-07-06 #3432 @PeerHerholz
Ahoi hoi @osorensen,

thx for the tag. I would be very interested in reviewing this important addition to the `python-neuroimaging` software stack. However, only if the folks involved think that I would be a fitting reviewer of course.

Cheers, Peer
