### 2021-07-08 #3432 @osorensen
@nicksisco1932, @robbisg, @DrMarc, would any of you be interested in reviewing this submission to the Journal of Open Source Software?
