### 2021-07-12 poldracklab/nitransforms#118 @PeerHerholz
JOSS review - current state & integration to NiBabel
Ahoi hoi folks,

this is in reference to the review I'm currently conducting for JOSS ([here](https://github.com/openjournals/joss-reviews/issues/3459)).

First of all, thank you very much for your dedicated work on this tool that definitely is much needed and fills an important gap in the `python neuroimaging` realm.  One central question I was wondering about and targets a lot of checkboxes in the review process is the current state & planned integration into `NiBabel` (#110 here & [#656 in NiBabel](https://github.com/nipy/nibabel/pull/656)). I understand that the general plan is to integrate `NiTransforms` as a `module`  into `NiBabel`. However, `NiTransforms`' current state appears a bit nebulous concerning its place and documentation. Regarding the first, it might be helpful to clearly state that `NiTransforms` evolved out of a planned addition to `Nibabel` to its own repository and package for development reasons, but upon finalization will be integrated back into `NiBabel`. This blends into the second point, documentation. While I assume that `functionality` and `API` will be part of the [NiBabel docs](https://nipy.org/nibabel/) after the planned `merge` (with `installation` instructions then becoming obsolete), it would still be necessary to have these aspects/resources during the transition period, given the review process and to allow interested parties to use the software. I saw a [docs folder](https://github.com/poldracklab/nitransforms/tree/master/docs) and [docs related PRs](https://github.com/poldracklab/nitransforms/commit/96250dfcc723d9938eea8f55cc0b790304e231a1), but wasn't able to find a respective link. Would it be possible to provide a link for the documentation and maybe update the [README](https://github.com/poldracklab/nitransforms#readme) to include the corresponding points?

I'm very sorry if I missed something obvious, misunderstood something or assumed something wrong. Please let me know if that is the case, as well as if you have any questions. Thanks again for creating this important software tool.

Cheers, Peer
