### 2021-07-13 poldracklab/nitransforms#118 @mgxd
Hi Peer!

You are right, docs are not being advertised! I added the link as the repo's website (https://nitransforms.readthedocs.io), but when browsing through noticed the API section is not populating. I'll take a look into that now.

@oesteban can you add me as a maintainer on RTD?
