### 2021-07-13 poldracklab/nitransforms#119 @PeerHerholz
JOSS review - software paper sections
Ahoi hoi folks,

this is in reference to the review I'm currently conducting for JOSS ([here](https://github.com/openjournals/joss-reviews/issues/3459)).

First of all, thank you very much for your dedicated work on this tool that definitely is much needed and fills an important gap in the `python neuroimaging` realm.  I enjoyed reading through your manuscript a lot and think it's already well-written, as well as very informative. Based on the [review process and included checkboxes](https://github.com/openjournals/joss-reviews/issues/3459#issue-939624504), I would like to suggest a few things.
The `summary` section could be split into a `summary` and a `statement of need` section, where the first would present the existing high-level overview (potentially in a slightly longer version) and the second would present the first half of the current `summary` section plus maybe a bit more on the limitations of the other software packages (but maybe also outline more: state-of-the-art workflows/pipelines do not necessarily operate within a single software package but combine multiple of them -> thus cross-software compatibility is needed).
The `spatial transforms` section is great and presents a very good overview on the topic, as well as existing problems. IMHO it should definitely be kept. However, it could potentially be a bit more "integrated" into the manuscript. Right now, at least for me, the section feels somewhat independent and slightly breaks the flow. I don't think much is needed, maybe 1-2 sentences at the beginning and end to situate the section more clearly in the manuscript and bridge from/to the other sections.

I'm very sorry if I missed something obvious, misunderstood something or assumed something wrong. Please let me know if that is the case, as well as if you have any questions. Thanks again for creating this important software tool.

Cheers, Peer
