### 2021-07-13 #3459 @PeerHerholz
To the authors,

I sincerely hope you and your loved ones are still doing ok during these ongoing difficult times.

At first, I would like to thank all authors for their very interesting submission and hard, dedicated work they put into it. The topic and aim of this resource are both fascinating and important. Within the following I would like to provide additional information regarding my review decision, which will be divided into the following respective issues (also linked above) opened within the [tool's repository](https://github.com/poldracklab/nitransforms):

- [current state & integration to NiBabel](https://github.com/poldracklab/nitransforms/issues/118)
- [software paper sections](https://github.com/poldracklab/nitransforms/issues/119)

In case some of my comments are misleading, difficult to understand or simply wrong based on an insufficient understanding on my side, I would like to apologize and kindly ask the authors to address such concerns. The same holds true for the nature of my comments: I aim to only provide helpful and constructive comments and criticism. If my wording appears too harsh or not helpful, please let me know and I will try my best to address this.

The presented tool, `NiTransforms`, allows researchers to work with an extensive amount and variety of `transform` formats. This includes both basic file handling (read/write/load/save), conversion between transform formats and application of transforms to images. Given the current state-of-the-art and increasing amount of workflows/pipelines that combine multiple software packages, the presented tool fills a prominent gap in the `python neuroimaging` ecosystem and will tremendously help to boost the reproducibility of neuroimaging results, while at the same time presenting the foundation for various other applications that could build upon it. Thus, it's clear "accept" from my side after the above outlined points were addressed/discussed.

Thanks again for this great contribution and please let me know if there are questions!

Cheers, Peer
