### 2021-07-13 poldracklab/nitransforms#118 @PeerHerholz
Hi @mgxd,

thank you very much for your quick reply and the information!
Ok, coolio, sounds great. Would you mind letting me know when you finished that work and I should have a look at it?

Thanks again.

Cheers, Peer
