+++ docs/_api/patched.rst
@@ -1,0 +1,6 @@
+=========================
+Patched Nibabel Functions
+=========================
+
+.. automodule:: nitransforms.patched
+    :members:
