+++ docs/_api/base.rst
@@ -1,0 +1,6 @@
+====
+Base
+====
+
+.. automodule:: nitransforms.base
+    :members:
