+++ docs/_api/linear.rst
@@ -1,0 +1,6 @@
+=================
+Linear Transforms
+=================
+
+.. automodule:: nitransforms.linear
+    :members:
