### 2021-07-13 d3d8ff8
DOC: Add documentation badge to README [skip ci]
