+++ README.md
@@ -4,6 +4,7 @@
 [![codecov](https://codecov.io/gh/poldracklab/nitransforms/branch/master/graph/badge.svg)](https://codecov.io/gh/poldracklab/nitransforms)
 [![doi](https://img.shields.io/badge/doi-10.31219%2Fosf.io%2F8aq7b-blue.svg)](https://doi.org/10.31219/osf.io/8aq7b)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/poldracklab/nitransforms/master?filepath=docs%2Fnotebooks%2F)
+[![Docs](https://readthedocs.org/projects/nitransforms/badge/?version=latest)](http://nitransforms.readthedocs.io/en/latest/?badge=latest)
 
 A development repo for [nipy/nibabel#656](https://github.com/nipy/nibabel/pull/656)
 
@@ -23,6 +24,6 @@ to apply transforms generated by the most popular neuroimaging packages and libr
 (AFNI, FSL, FreeSurfer, ITK, and SPM).
 The proposed software is subject to continuous integration tests to check the
 compatibility with each supported tool after every change to the code base.
-Compatibility between software tools and imaging formats is a necessary bridge 
+Compatibility between software tools and imaging formats is a necessary bridge
 to ensure the reproducibility of results and enable the optimization and evaluation
 of current image processing and analysis workflows.
