### 2021-07-13 poldracklab/nitransforms#118 @mgxd
Should be all set after #120 - https://nitransforms.readthedocs.io/en/latest/api.html
