### 2021-07-13 poldracklab/nitransforms#118 @oesteban
You're a maintainer now :)

On Tue, Jul 13, 2021, 21:16 Mathias Goncalves ***@***.***>
wrote:

> Should be all set after #120
> <https://github.com/poldracklab/nitransforms/pull/120> -
> https://nitransforms.readthedocs.io/en/latest/api.html
>
