+++ joss/nt.bib
@@ -1,14 +1,14 @@
 @article{cox_software_1997,
 	author = {Cox, Robert W. and Hyde, James S.},
 	copyright = {Copyright © 1997 John Wiley \& Sons, Ltd.},
-	doi = {10.1002/(SICI)1099-1492(199706/08)10:4/5{	extless}171::AID-NBM453{	extgreater}3.0.CO;2-L},
+	doi = {10.1002/(SICI)1099-1492(199706/08)10:4/5<171::AID-NBM453>3.0.CO;2-L},
 	issn = {1099-1492},
 	journal = {NMR Biomed},
 	language = {en},
 	number = {4-5},
 	pages = {171-178},
 	title = {Software tools for analysis and visualization of fMRI data},
-	url = {http://onlinelibrary.wiley.com/doi/10.1002/(SICI)1099-1492(199706/08)10:4/5{	extless}171::AID-NBM453{	extgreater}3.0.CO;2-L/abstract},
+	url = {http://onlinelibrary.wiley.com/doi/10.1002/(SICI)1099-1492(199706/08)10:4/5<171::AID-NBM453>3.0.CO;2-L},
 	volume = 10,
 	year = 1997
 }
