### 2021-07-14 poldracklab/nitransforms#118 @oesteban
> I understand that the general plan is to integrate `NiTransforms` as a `module` into `NiBabel`. However, `NiTransforms`' current state appears a bit nebulous concerning its place and documentation. Regarding the first, it might be helpful to clearly state that `NiTransforms` evolved out of a planned addition to `Nibabel` to its own repository and package for development reasons, but upon finalization will be integrated back into `NiBabel`.

Yes, this is a great point. *NiTransforms* was envisioned as part of *NiBabel* since the beginning. However, shortly after starting out as a PR to *NiBabel*, we realized that it was going to build up in a humongous PR nobody would be able to review as thoroughly as it would require. Also, *NiTransforms* has many connections to BIDS/BIDS-Derivatives and its X5 format specification for transforms, which falls outside of the current scope of *NiBabel*.

So the plan is to make it an isolated tool, and once it is finished, integrate with *NiBabel*. On its end, *NiBabel* just initiated its CZI EOSS, so there's quite a lot of work ahead refactoring the library. The grant also covers aspects of *NiTransforms*. Once *NiTransforms* is ready for integration, we will define what can go in *NiBabel* (presumably everything, except perhaps some final details of the X5 implementation, although *NiBabel* will support the data structure at least logically).

This is to say that the chances that *NiTransforms* is integrated into *NiBabel* are high and scheduled to happen in ~10 months. The "bus factor" is high in this case thanks to the CZI funding. However, we believe *NiTransforms* can be considered as a software component of its own that deserves a focused publication at JOSS.

Regarding documentation, there's a whole lot of action items in the CZI grant for NiBabel (and NiTransforms too). When the implementation of all planned features is concluded, we will start generating tutorials. Some of them may also be covered by other funds we are trying to put together around NiPreps (www.nipreps.org).

Finally, we are planning to transfer over the repo under nipy (or at least, nipreps). We haven't done so this far as the poldracklab organization uses fewer resources from GitHub Actions and CircleCI at the moment. However, if you believe this would be important for the visibility of the project or other reasons I cannot think of right now, we are happy to push the transfer sooner than expected.
