### 2021-07-14 poldracklab/nitransforms#118 @oesteban
> However, we believe _NiTransforms_ can be considered as a software component of its own that deserves a focused publication at JOSS.

Which doesn't mean we will not follow a long deprecation cycle for those who try the code as in the JOSS publication. Importing from nitransforms will be supported for a long time (as a shim that uses nibabel underneath) until we are sure everyone is using nibabel's import directly.
