### 2021-07-14 e6d23b9
FIX: Suggested edits to the JOSS submission
Updates the JOSS submission to incorporate the suggestions by @PeerHerholz.

References: openjournals/joss-reviews#3459
Resolves: #119.
