### 2021-07-14 1df315d
DOC: Add an _Installation_ section
