### 2021-07-14 poldracklab/nitransforms#123 @robbisg
JOSS review - software 
Hi,

this issue is related to JOSS submission (https://github.com/openjournals/joss-reviews/issues/3459) and focuses more on the software side.

I think that this package can be very useful to easily integrate different formats and algorithms into more complicated pipelines, but also with the adoption of BIDS specification, so thanks for sharing this package with the community.

I noticed that you added the documentation website with the API of the tools, you have a set of test with a very high coverage and automated CI, and it's great, since it takes lot of time to code these parts!

Some points needs to be added in my opinion (but also from JOSS guidelines :) ):
* __examples__: from a user point of view, as you know, this will help users to understand and use (correctly) your package.
* __installation guidelines__: I suppose that the good old `pip install` or `python setup.py install` are used to install the package, but this should be mentioned in the repo README with dependencies.
* __contribution guidelines__: this is a minor edit since you planned to merge this package into `nibabel` which have this section, but a couple of lines in the docs / README are needed until this package will be independent and for JOSS, as well.

I saw that some of these concerns have been answered in #118 .

I have an extra question, that is more a curiosity than a issue related to the tool: do you think it would be useful to include in your package some atlas transformations utilities (e.g. `to_talairach`, `to_mni`) or is out of the scope of it?

Cheers,
Roberto

PS: I saw while I was writing that some of these issues have been addressed by #122
