### 2021-07-14 poldracklab/nitransforms#123 @robbisg
## Paper review

I was opening another issue related to paper edits, but I saw that in #121 you just edited the paper and addressed lot of stuff there.

I have an extra suggestion: a comparison with other tools that are already available is needed for JOSS, and I think can help to understand the importance of your tool. So you can add a sentence if yours is the first the tool that address the conversion of transformation file formats, or add a short comparison with other tools.
