### 2021-07-14 #3459 @robbisg
The proposed tool, `nitransforms`, is used to convert transformation files and manage them very easily.

I think that this tool is very useful and make it available will extremely help the python neuroimaging community. I agree with @PeerHerholz that this tool is very useful to ease building complex and automated pipelines and help the reproducibility of neuroimaging studies.

I opened an issue in the package repository (https://github.com/poldracklab/nitransforms/issues/123) with my suggestions.
My suggestions are related to the documentation part, I encouraged the authors to add examples or tutorials, since from my perspective, this is the entry point of new users and a way to understand the usefulness of the software.
Moreover a brief overview of other tools used for this purposes should be included in the paper, if no other tools are available, a clear statement that this is the only tool with these goals can help to add importance to it.

Once that these suggestions will be addressed, I will be happy to accept the publication to JOSS.

If you have any other questions, let me know. I would like to apologize if some of my comments are misleading.

Cheers,
Roberto
