### 2021-07-14 poldracklab/nitransforms#123 @oesteban
> very high coverage and automated CI, and it's great, since it takes lot of time to code these parts!

This was part of the vision - the reliability our implementation requires is only possible to achieve under test-driven development approaches.

> * **examples**: from a user point of view, as you know, this will help users to understand and use (correctly) your package.

Thanks - yes this is a great suggestion. We already have some notebooks, I'll see how to integrate them into the docs ASAP. We can also add a link to https://www.nipreps.org/nipreps-book/tutorial/registration.html, where the API is demonstrated in a real scenario. Finally, we have a binder link - the container is pretty heavy so it fails often, but it is worth adding somewhere and optimize the container down the line.

> that is more a curiosity than a issue related to the tool: do you think it would be useful to include in your package some atlas transformations utilities (e.g. `to_talairach`, `to_mni`) or is out of the scope of it?

I think this would be the responsibility of downstream software as _NiTransforms_ is completely unaware of the semantics of spaces.

_NiTransforms_ _only_ represents spatial transforms, so if a user wants some `to_talairach()` function/member of some object, then they will be able to leverage the API with:
1. Load a spatial transform, as generated by the software that was used to register the target image into Talairach space.
2. Set some reference in the Talairach space (typically a template, check efforts on www.templateflow.org to complete this)
3. Call the `.apply()` API of the transform object.

> * **contribution guidelines**: this is a minor edit since you planned to merge this package into `nibabel` which have this section, but a couple of lines in the docs / README are needed until this package will be independent and for JOSS, as well.

Sure, it doesn't hurt to say we abide by the contribution guidelines of nibabel and all that applies to nibabel applies here. Thanks!

> a comparison with other tools that are already available is needed for JOSS, and I think can help to understand the importance of your tool. So you can add a sentence if yours is the first the tool that address the conversion of transformation file formats, or add a short comparison with other tools.

A tricky but important one. I guess we can consider here some converters inside some software (e.g., FreeSurfer's conversion between LTA and FSL formats) and more generally C3D (which goal is conversion indeed). I'll chat with @mgxd to improve along these lines.
