### 2021-07-14 poldracklab/nitransforms#118 @PeerHerholz
Hi @oesteban,

thank you very much for your response and the additional information.
IMHO the plan/ideas outlined by you are very reasonable and understandable. I completely agree with `NiTransforms`' role as a dedicated software package/tool that definitely deserves its own publication. My comments were rather targeted at including this information in the `docs` and potentially manuscript. I'm very sorry for this misunderstanding.

P.S.: The entire work package around `NiBabel`'s CZI seems fantastic!
