### 2021-07-14 poldracklab/nitransforms#118 @PeerHerholz
Hi @mgxd (& @oesteban),

thanks a lot for updating the `docs` and adding the link to the repo.
Would it be possible to discuss a few further things? All the points I would like to bring up are actually perfectly summarized by @robbisg in the respective issue: `installation`, `examples` and `contribution guidelines`. Thus, I'll just point to the comments outlined there. One small general addition that would be cool: a `CoC` and one potential addition to the `docs`: the `spatial transforms` section of the paper, as it would add some great background information there as well.
