### 2021-07-15 cefd517
DOC: Add installation verification
