### 2021-07-15 a3e43e9
DOC: Add example notebooks to sphinx documentation
