+++ docs/examples.rst
@@ -1,0 +1,12 @@
+========
+Examples
+========
+
+A collection of Jupyter Notebooks to serve as interactive tutorials.
+
+.. toctree::
+    :maxdepth: 1
+
+    notebooks/01_preparing_images
+    notebooks/02_afni_deoblique
+    notebooks/isbi2020
