### 2021-07-19 9cb8c24
doc: add CONTRIBUTING.md file pointing at NiBabel
