+++ .github/workflows/CONTRIBUTING.md
@@ -1,0 +1,16 @@
+# Contributing to *NiTransforms* (a *NiBabel* feature-repo)
+
+Welcome to *NiBabel*, and the *NiTransforms* repository!
+We're excited you're here and want to contribute.
+
+Please see the [NiBabel Developer Guidelines][link_devguide] on our
+on our [documentation website][link_docs].
+
+These guidelines are designed to make it as easy as possible to get involved.
+If you have any questions that aren't discussed in our documentation, or it's
+difficult to find what you're looking for, please let us know by opening an
+[issue][link_issues]!
+
+[link_docs]: https://nipy.org/nibabel
+[link_devguide]: https://nipy.org/nibabel/devel/devguide.html
+[link_issues]: https://github.com/poldracklab/fmriprep/issues
