### 2021-07-19 8790bbe
doc: clarify integrations plans with nibabel
