### 2021-07-19 3370364
FIX: Misinterpretation of voxel ordering in LTAs
The structarray was used directly and the extra axis actually changed
the order of operations when the direction cosines were scaled by the
voxel sizes.

A new test case has been added for which this error was apparent.
This bug caused nipreps/fmriprep#2307, nipreps/fmriprep#2393, and
nipreps/fmriprep#2410.
nipreps/fmriprep#2444 just works around the problem by using
``lta_convert`` instead of *NiTransforms*.
The ``lta_convert`` tool can be now dropped.

Resolves: #125
