### 2021-07-20 1d1d202
enh: add test, address @effigies' comments from review
Co-authored-by: Chris Markiewicz <effigies@gmail.com>
