+++ README.md
@@ -35,4 +35,4 @@ Also, *NiTransforms* has many connections to BIDS/BIDS-Derivatives and its X5 fo
 
 The plan is to make it an isolated tool, and once it is finished, proceed with the integration into *NiBabel*.
 Once this repository is ready for integration, we will define what can go into *NiBabel* (presumably everything, except perhaps some final details of the X5 implementation, although *NiBabel* will support the data structure at least logically).
-This is to say that the chances that NiTransforms is integrated into NiBabel are high and scheduled to happen in ~10 months.
+This is to say that the chances that NiTransforms is integrated into NiBabel are high and scheduled to happen in ~2022 Q2.
