### 2021-07-21 c1fd4ef
ENH: Add one test for the implementation of volume geometry of LTA
The new LTA file and the matrix hard-coded on the test come from an
issue discovered in *fMRIPrep* (see nipreps/fmriprep#2444).

This test currently breaks master.
