### 2021-07-21 d2defb0
enh: @effigies' suggestions on code review
Co-authored-by: Chris Markiewicz <effigies@gmail.com>
