### 2021-07-21 0ebf0e6
Update nitransforms/io/lta.py
Co-authored-by: Chris Markiewicz <markiewicz@stanford.edu>
