### 2021-08-09 38e0407
DOC: Add contributing page to RTD
