+++ README.md
@@ -33,6 +33,6 @@ of current image processing and analysis workflows.
 Shortly after starting with [nipy/nibabel#656](https://github.com/nipy/nibabel/pull/656), it became apparent that it was going to build up in a humongous PR nobody would be able to review as thoroughly as it would require.
 Also, *NiTransforms* has many connections to BIDS/BIDS-Derivatives and its X5 format specification for transforms, which falls outside of the current scope of *NiBabel*.
 
-The plan is to make it an isolated tool, and once it is finished, proceed with the integration into *NiBabel*.
+The plan is to make it an isolated tool, and once it is pertinent, proceed with the integration into *NiBabel*.
 Once this repository is ready for integration, we will define what can go into *NiBabel* (presumably everything, except perhaps some final details of the X5 implementation, although *NiBabel* will support the data structure at least logically).
 This is to say that the chances that NiTransforms is integrated into NiBabel are high and scheduled to happen in ~2022 Q2.
