### 2021-08-21 #3459 @PeerHerholz
Ahoi hoi folks,

thx so much for all the work on this @mgxd & @oesteban, I think the changes look great!

Two small things: as the contribution guidelines were added, I was wondering about the `CoC` and regarding the examples, I wanted to ask if a few sentences/explanations could be added to the [Preparing Images](https://nitransforms.readthedocs.io/en/latest/notebooks/01_preparing_images.html) and [AFNI Deoblique](https://nitransforms.readthedocs.io/en/latest/notebooks/02_afni_deoblique.html) notebooks, as well as functionality to download the examples either as `.ipynb` or `.py`?

Thanks again, cheers, Peer
