### 2021-08-21 #3459 @robbisg
Hi everyone,

sorry for the late reply.
Thank you for your changes.
I agree with @PeerHerholz that examples need a few lines of description of what's going on in the script.
I think that a brief overview (one sentence) of similar package or about the lack of tools like yours can be added to the paper.

For the rest, you did a very good job! ;)

Cheers,
Roberto
