### 2021-09-08 002f4c0
DOC: Transfer CoC from NiBabel
