### 2021-09-08 f8f634d
fix: add mention to potential alternatives
