+++ joss/nt.bib
@@ -73,3 +73,10 @@
 	title = {Statistical parametric mapping : the analysis of functional brain images},
 	year = 2006
 }
+
+@article{yushkevich_open_nodate,
+	title = {Open {Source} {Software}: {Convert3D} {Medical} {Imaging} {Processing} tool},
+	url = {https://sourceforge.net/p/c3d},
+	journal = {Sourceforge},
+	author = {Yushkevich, Paul A.},
+}
