### 2021-09-09 #3459 @oesteban
@PeerHerholz @robbisg @osorensen thanks much for your patience.

We've pushed the following changes:

> @PeerHerholz
> *I was wondering about the `CoC`*

https://github.com/poldracklab/nitransforms/commit/be5341b5f37b83b25e398799908bc65ba4bf4b7b

> @PeerHerholz
> *a few sentences/explanations could be added to the Preparing Images and AFNI Deoblique notebooks, as well as functionality to download the examples either as `.ipynb` or `.py`?*
> @robbisg
> *examples need a few lines of description of what's going on in the script.*

We have added the due descriptions and also included an automated prologue of the nbsphinx extension to include links to the source code of the examples:
  * https://nitransforms.readthedocs.io/en/latest/notebooks/01_preparing_images.html
  * https://nitransforms.readthedocs.io/en/latest/notebooks/02_afni_deoblique.html
  * https://nitransforms.readthedocs.io/en/latest/notebooks/isbi2020.html

> @robbisg
> *a brief overview (one sentence) of similar package or about the lack of tools like yours can be added to the paper.*

[poldracklab/nitransforms#132 (files)](https://github.com/poldracklab/nitransforms/pull/132/files)

Please let us know if there's anything else we might have missed in addressing all the comments. Thanks very much for the effort and time you've sunk into this, it's been really worthwhile on our side for improving the accessibility and usability of the project and the JOSS submission itself. Very much appreciated.
