### 2021-09-09 #3459 @PeerHerholz
Ahoi hoi folks,

thx so much for all the work @oesteban & @mgxd, the changes look great!

IMHO the [AFNI example](https://nitransforms.readthedocs.io/en/latest/notebooks/02_afni_deoblique.html) could still use a bit more explanations, however, I think moving the comments out of the larger code blocks might already be enough. Sorry, I really don't mean to be nitpicking, but just trying to approach the materials from a potential user perspective. I would check my remaining box asap after that!

Cheers, Peer
