### 2021-09-09 #3459 @oesteban
@PeerHerholz - what about https://github.com/poldracklab/nitransforms/pull/134 ? Do you think that would clarify things up?
