### 2021-09-09 #3459 @robbisg
Hi,

I agree with @PeerHerholz ! it is one of the first things that I visit when a new tool is released.

Given your brand new fix, you can split also cell [79] in different titled cells, from my side.

The paragraph about other tools is ok! :)

Thank you for your work!
Roberto
