### 2021-09-09 #3459 @PeerHerholz
Ahoi hoi @oesteban,

damn that was fast! Yeah, looks great and comprehensible, thanks for adding the explanations.
I'll check my last box and thus the submission is good to go from my end (@osorensen). Thanks again to the authors for all their great work and this important addition to the python-neuroimaging-world, I can't wait to work with `NiTransforms`!

Cheers, Peer
