### 2021-09-09 #3459 @osorensen
@oesteban, could you please fix the following issues in the paper?

- Add a reference to Figure 1 in the text. Maybe somewhere in the Methods section on page 2?
- Page 3: Add caption to Figure 1.
- Page 3: In Part (1) of Figure 1, should "... in the moving's coordinate system ..." be corrected to "... in the moving image's coordinate system ..."?
