### 2021-09-10 ca4d450
fix: final edits to JOSS submission
Edits to the submission text following @osorensen's suggestions (https://github.com/openjournals/joss-reviews/issues/3459#issuecomment-916088981):
* Add a reference to Figure 1 in the text. Maybe somewhere in the Methods section on page 2?
* Page 3: Add caption to Figure 1.
