### 2021-09-10 d44112a
docs: add fMRIPrep reference into JOSS submission, address minimal problems
