+++ joss/nt.bib
@@ -80,3 +80,21 @@
 	journal = {Sourceforge},
 	author = {Yushkevich, Paul A.},
 }
+
+@article{esteban_fmriprep_2019,
+	title = {{fMRIPrep}: a robust preprocessing pipeline for functional {MRI}},
+	volume = {16},
+	copyright = {2018 The Author(s), under exclusive licence to Springer Nature America, Inc.},
+	issn = {1548-7105},
+	shorttitle = {{fMRIPrep}},
+	url = {https://www.nature.com/articles/s41592-018-0235-4},
+	doi = {10.1038/s41592-018-0235-4},
+	language = {En},
+	number = {1},
+	urldate = {2018-12-10},
+	journal = {Nat Meth},
+	author = {Esteban, Oscar and Markiewicz, Christopher J. and Blair, Ross W. and Moodie, Craig A. and Isik, A. Ilkay and Erramuzpe, Asier and Kent, James D. and Goncalves, Mathias and DuPre, Elizabeth and Snyder, Madeleine and Oya, Hiroyuki and Ghosh, Satrajit S. and Wright, Jessey and Durnez, Joke and Poldrack, Russell A. and Gorgolewski, Krzysztof J.},
+	month = jan,
+	year = {2019},
+	pages = {111--116},
+}
