### 2021-09-10 #3459 @oesteban
Sure thing, I just realized a self-citation was missed (fMRIPrep was the only piece of software mentioned without a due reference). I'll check on the Zenodo record - I believe we set it up, but I might be wrong. Anyways that was on the todo list down the line, so happy to get it right before finishing with the review stage.
