### 2021-09-10 #3459 @osorensen
Good! I see there is a Zenodo archive for the paper (https://osf.io/8aq7b/), but please create one for the final version of the software too, so that we can get a doi link for the version that gets published.
